Fix stray stderr output on machines with GNU which
On a NixOS machine I was getting a message like the following on every
Rails boot:

> which: no bun in (... [my path] ...)

I chased it down to this library's usage of `which` to find the mjml
binary. Using backticks to call it meant stderr was passing through to
the terminal.

Ubuntu and MacOS use different versions of `which` that do not
print to stderr, so it's likely not many people have run into this
issue.